Revises #13
The code has drifted away, so it's impossible to know what exactly
caused the problem. It might have been some combination of testing
website and Rakudo changes, but this problem no longer exists, so this
closes #6.

Also, now http redirects to https, so eliminates a few tests that
actually do the same.